# models/machine_output.go
package models

import (
	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

// MachineOutput Machine Output
// swagger:model machine-output
type MachineOutput struct {
	MachineInput
}

// UnmarshalJSON unmarshals this object from a JSON structure
func (m *MachineOutput) UnmarshalJSON(raw []byte) error {

	var aO0 MachineInput
	if err := swag.ReadJSON(raw, &aO0); err != nil {
		return err
	}
	m.MachineInput = aO0

	return nil
}

// MarshalJSON marshals this object to a JSON structure
func (m MachineOutput) MarshalJSON() ([]byte, error) {
	var _parts [][]byte

	aO0, err := swag.WriteJSON(m.MachineInput)
	if err != nil {
		return nil, err
	}
	_parts = append(_parts, aO0)

	return swag.ConcatJSON(_parts...), nil
}

// Validate validates this machine output
func (m *MachineOutput) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.MachineInput.Validate(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}