# swagger.yaml
swagger: '2.0'
info:
  version: v3.0
  title: rocket-skates
  description: Protocol for Bare Metal Provider
host: 'localhost:8092'
basePath: /api/v3
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  rebar:
    type: basic
paths:
...
  /machines:
    get:
      operationId: list-machines
      summary: List Machines
      tags:
        - Machines
      responses:
        '200':
          description: ''
          schema:
            type: object
            properties:
              Result:
                $ref: '#/definitions/result'
              Data:
                type: array
                items:
                  $ref: '#/definitions/machine-output'
        '401':
          $ref: '#/responses/trait:commonGetResponses:401'
        '404':
          $ref: '#/responses/trait:commonGetResponses:404'
        '500':
          $ref: '#/responses/trait:commonGetResponses:500'
    post:
      operationId: post-machine
      summary: Create Machine
      tags:
        - Machines
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/machine-input'
            example:
              Address: 192.168.124.84
              BootEnv: local
              Name: d52-54-3d-49-00-00.local.neode.org
              Params: {}
              TenantId: 1
              Uuid: 0f2a8f98-af51-415e-ad4f-cfea69adaed3
      responses:
        '201':
          description: ''
          schema:
            type: object
            properties:
              Result:
                $ref: '#/definitions/result'
              Data:
                $ref: '#/definitions/machine-output'
        '401':
          $ref: '#/responses/trait:commonPostResponses:401'
        '409':
          $ref: '#/responses/trait:commonPostResponses:409'
        '500':
          $ref: '#/responses/trait:commonPostResponses:500'
  '/machines/{uuid}':
    parameters:
      - name: uuid
        in: path
        required: true
        type: string
    get:
      operationId: get-machine
      summary: Get Machine
      tags:
        - Machines
      responses:
        '200':
          description: ''
          schema:
            type: object
            properties:
              Data:
                $ref: '#/definitions/machine-output'
              Result:
                $ref: '#/definitions/result'
        '401':
          $ref: '#/responses/trait:commonGetResponses:401'
        '404':
          $ref: '#/responses/trait:commonGetResponses:404'
        '500':
          $ref: '#/responses/trait:commonGetResponses:500'
    put:
      operationId: put-machine
      summary: Update Machine
      tags:
        - Machines
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/machine-input'
            example:
              Address: 192.168.124.84
              BootEnv: local
              Name: d52-54-3d-49-00-00.local.neode.org
              Params: {}
              TenantId: 1
              Uuid: 0f2a8f98-af51-415e-ad4f-cfea69adaed3
      responses:
        '200':
          description: ''
          schema:
            type: object
            properties:
              Result:
                $ref: '#/definitions/result'
              Data:
                $ref: '#/definitions/machine-output'
        '401':
          $ref: '#/responses/trait:commonPutResponses:401'
        '404':
          $ref: '#/responses/trait:commonPutResponses:404'
        '500':
          $ref: '#/responses/trait:commonPutResponses:500'
    patch:
      operationId: patch-machine
      summary: Patch Machine
      tags:
        - Machines
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/patchinput'
      responses:
        '202':
          description: ''
          schema:
            type: object
            properties:
              Result:
                $ref: '#/definitions/result'
              Data:
                $ref: '#/definitions/machine-output'
        '401':
          $ref: '#/responses/trait:commonPatchResponses:401'
        '404':
          $ref: '#/responses/trait:commonPatchResponses:404'
        '417':
          $ref: '#/responses/trait:commonPatchResponses:417'
        '500':
          $ref: '#/responses/trait:commonPatchResponses:500'
    delete:
      operationId: delete-machine
      summary: Delete Machine
      tags:
        - Machines
      responses:
        '204':
          description: ''
        '404':
          $ref: '#/responses/trait:commonDeleteResponses:404'
        '409':
          $ref: '#/responses/trait:commonDeleteResponses:409'
        '500':
          $ref: '#/responses/trait:commonDeleteResponses:500'
responses:
  'trait:commonGetResponses:401':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonGetResponses:404':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonGetResponses:500':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonDeleteResponses:404':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonDeleteResponses:409':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonDeleteResponses:500':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPostResponses:401':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPostResponses:409':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPostResponses:500':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPutResponses:401':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPutResponses:404':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPutResponses:500':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPatchResponses:401':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPatchResponses:404':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPatchResponses:417':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPatchResponses:500':
    description: ''
    schema:
      $ref: '#/definitions/result'
definitions:
  machine-input:
    title: Machine Input
    type: object
    properties:
      Address:
        type: string
        format: ipv4
      BootEnv:
        type: string
      Name:
        type: string
        format: hostname
      Params:
        $ref: '#/definitions/parameterobjectmap'
      TenantId:
        type: integer
      Uuid:
        type: string
        format: uuid
    example:
      Address: 192.168.124.84
      BootEnv: local
      Name: d52-54-3d-49-00-00.local.neode.org
      Params: {}
      TenantId: 1
      Uuid: 0f2a8f98-af51-415e-ad4f-cfea69adaed3
  machine-output:
    title: Machine Output
    allOf:
      - $ref: '#/definitions/machine-input'
  parameterobjectmap:
    title: ParameterObjectMap
    AdditionalProperties:
      type: object
  result:
    title: Result
    type: object
    properties:
      Code:
        type: integer
      Messages:
        type: array
        items:
          type: string
...
  patchinput:
    title: PatchInput
    type: array
    items:
      $ref: '#/definitions/patchline'
  patchline:
    title: PatchLine
    type: object
    properties:
      Op:
        type: string
      Path:
        type: string
      From:
        type: string
      Value: {}
...